refactor: Remove `ModbusClient` related classes
