--- test/ascii-edge-cases.test.ts
@@ -1,129 +1,0 @@
-// Test file to validate ASCII edge cases and identify any gaps
-import { describe, expect, it } from "vitest";
-import { calculateLRC } from "../src/lrc.ts";
-import { ModbusClient } from "../src/modbus.ts";
-
-describe("ASCII Edge Cases Analysis", () => {
-  it("handles invalid hex characters", async () => {
-    const client = new ModbusClient();
-    client.protocol = "ascii";
-
-    const errorSpy: Error[] = [];
-    client.on("error", (error) => errorSpy.push(error));
-
-    const promise = client.read({
-      functionCode: 3,
-      quantity: 1,
-      slaveId: 1,
-      startAddress: 0,
-    });
-
-    // Frame with invalid hex character 'G'
-    const responseFrame = ":01030G000AF0\r\n";
-    client.handleResponse(
-      new Uint8Array(Array.from(responseFrame).map((c) => c.charCodeAt(0))),
-    );
-
-    await expect(promise).rejects.toThrow(/Invalid hex pair/);
-  });
-
-  it("handles noise before frame start", async () => {
-    const client = new ModbusClient();
-    client.protocol = "ascii";
-
-    const promise = client.read({
-      functionCode: 3,
-      quantity: 1,
-      slaveId: 1,
-      startAddress: 0,
-    });
-
-    // Noise followed by valid frame
-    const responseFrame = "NOISE123ABC:010302000AF0\r\n";
-    client.handleResponse(
-      new Uint8Array(Array.from(responseFrame).map((c) => c.charCodeAt(0))),
-    );
-
-    const response = await promise;
-    expect(response.data).toEqual([10]);
-  });
-
-  it("allows caller to abort incomplete frame (no implicit timeout)", async () => {
-    const client = new ModbusClient();
-    client.protocol = "ascii";
-    const controller = new AbortController();
-    const promise = client.read(
-      { functionCode: 3, quantity: 1, slaveId: 1, startAddress: 0 },
-      { signal: controller.signal },
-    );
-    const responseFrame = ":010302000AF0"; // missing CRLF so never completes
-    client.handleResponse(
-      new Uint8Array(Array.from(responseFrame).map((c) => c.charCodeAt(0))),
-    );
-    controller.abort();
-    await expect(promise).rejects.toThrow(/aborted/i);
-  });
-
-  it("handles odd number of hex characters", async () => {
-    const client = new ModbusClient();
-    client.protocol = "ascii";
-
-    const promise = client.read({
-      functionCode: 3,
-      quantity: 1,
-      slaveId: 1,
-      startAddress: 0,
-    });
-
-    // Frame with odd number of hex chars
-    const responseFrame = ":01030200AF0\r\n"; // Missing one char
-    client.handleResponse(
-      new Uint8Array(Array.from(responseFrame).map((c) => c.charCodeAt(0))),
-    );
-
-    await expect(promise).rejects.toThrow(/odd number of hex characters/);
-  });
-
-  it("handles frame too short", async () => {
-    const client = new ModbusClient();
-    client.protocol = "ascii";
-
-    const promise = client.read({
-      functionCode: 3,
-      quantity: 1,
-      slaveId: 1,
-      startAddress: 0,
-    });
-
-    // Frame too short (less than 3 bytes)
-    const responseFrame = ":0103\r\n";
-    client.handleResponse(
-      new Uint8Array(Array.from(responseFrame).map((c) => c.charCodeAt(0))),
-    );
-
-    await expect(promise).rejects.toThrow(/ASCII frame too short/);
-  });
-
-  it("validates LRC calculation matches expected", () => {
-    // Test the specific LRC calculation from the test
-    const messageBytes = [0x01, 0x03, 0x02, 0x00, 0x0a]; // from :010302000AF0
-    const calculatedLRC = calculateLRC(messageBytes);
-    expect(calculatedLRC).toBe(0xf0);
-  });
-
-  it("caller can abort when only noise is received", async () => {
-    const client = new ModbusClient();
-    client.protocol = "ascii";
-    const controller = new AbortController();
-    const promise = client.read(
-      { functionCode: 3, quantity: 1, slaveId: 1, startAddress: 0 },
-      { signal: controller.signal },
-    );
-    const responseFrame = "ABCDEFGHIJ123456"; // noise
-    client.handleResponse(
-      new Uint8Array(Array.from(responseFrame).map((c) => c.charCodeAt(0))),
-    );
-    controller.abort(new Error("Aborted"));
-    await expect(promise).rejects.toThrow(/aborted/i);
-  });
-});
