+++ test/frameBuilder.test.ts
@@ -1,6 +1,9 @@
 import { describe, expect, it } from "vitest";
 import { buildReadRequest, buildWriteRequest } from "../src/frameBuilder.ts";
-import type { ModbusReadConfig, ModbusWriteConfig } from "../src/modbus.ts";
+import type {
+  ModbusReadConfig,
+  ModbusWriteConfig,
+} from "../src/types/modbus.ts";
 
 describe("Frame Builder", () => {
   describe("buildReadRequest", () => {
